+++ packer/communicator/powershell/driver_ps_4.go
@@ -2,7 +2,7 @@
 // All Rights Reserved.
 // Licensed under the Apache License, Version 2.0.
 // See License.txt in the project root for license information.
-package driver
+package powershell
 
 import (
 	"bytes"
